# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Forever Medspa",
        "description": "All API Related to Giftcards",
        "termsOfService": "test",
        "contact": {
            "name": "Forever Medspa",
            "url": "https://forevermedspanj.com",
            "email": "info@forevermedspanj.com"
        },
        "license": {
            "name": "License",
            "url": "https://forevermedspanj.com/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/forever-medspa-giftcard/public/api/"
        }
    ],
    "paths": {
...
        "/category/{id}": {
            "get": {
                "tags": [
                    "Category Management"
                ],
                "summary": "View Category Id Wise",
                "description": "View the list of product categories.",
                "operationId": "3d326ac6f476a643ab8099f802390db3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "user_token",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "FOREVER-MEDSPA"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Product category found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found: No product category found"
                    }
                }
            }
        },
...
        "/product/{id}": {
            "get": {
                "tags": [
                    "Product Management"
                ],
                "summary": "View Product Id Wise",
                "description": "View the list of product categories.",
                "operationId": "93a7bd083b4cfd013718feed46ccb752",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "user_token",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "FOREVER-MEDSPA"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Product found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found: No product found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}